drbd: Move clearing the CRASHED_PRIMARY bit into finish_state_change()
Up to now it was in w_after_state_chage(). That was misplaced. Since the
after state change function might execute considerably later than when
the state change finishes.
Clearing it with the state change can avoid various bad effects on
other things, like sending it set during a handshake, and clearing it
before the local UUID comparison happens.