drbd: fix depmod.d override for rhel and derivatives
SuSE is not affected, because their default search path is
    search updates extra weak-updates kgraft built-in
Red Hat has
    search updates extra built-in weak-updates

Actulally affected is likely only oracle linux,
rhel does not ship the "in-tree" drbd module.

Oracle linux does, so modutils would find the installed
drbd 9.x only if it matches the exact kernel version
(in /lib/modules/<kernel-version>/extra/drbd/drbd.ko),
but not for a "weak-updates" compatible build for a
slightly different kernel version
(in /lib/modules/<kernel-version>/weak-updates/drbd/drbd.ko,
symlinked to /lib/modules/<other-kernel-version>/extra/drbd/drbd.ko).

Even if you have the compatible 9.x installed,
in that case it would use the in-tree "8.4-equivalent" module.

The override path is not a prefix-match, but needs to match the relative
sub-directory exactly, see is_higher_priority() in
https://git.kernel.org/pub/scm/utils/kernel/module-init-tools/module-init-tools.git/tree/depmod.c#n713

fix: use weak-updates/drbd

The weak-modules magic should not place those symlinks if it finds a matching
module already installed into "extras", but should that ever become a problem,
it would need to become:

   override drbd * weak-updates/drbd
   override drbd_transport_tcp * weak-updates/drbd
   override drbd * extra/drbd
   override drbd_transport_tcp * extra/drbd

(the one listed last is the one that is prioritized)

Alternatively, we could install as "/etc/depmod.d/00_drbd.conf",
and change the system wide search path. But that feels too hackish.