drbd: only send P_PEERS_IN_SYNC every 4MiB
The removal of the resync LRU and associated resync extent concept also
affects P_PEERS_IN_SYNC. This was previously sent whenever a resync
extent was cleared. We no longer have this structure to determine when
to send P_PEERS_IN_SYNC. Instead, send it when:
* Resync progresses a given amount
* Resync finishes
* P_PEERS_IN_SYNC is received and clears some bits

We do not need to send it when application activity clears bitmap bits
because those bits are already cleared and set in the write path
appropriately. There is never a need to send P_PEERS_IN_SYNC to a
further away peer as a result of bits being cleared in this way. If the
peer did not receive the write itself then we will always have bits set
in our bitmap for it.