drbd: fix use of __drbd_chk_io_error()
in the drbd-9.0 branch __drbd_chk_io_error() might be used where
the req_lock was hold. In the drbd-9.1 branch this lock was removed
and it happened that all call sites of __drbd_chk_io_error() only
hold a read-lock on state_rwlock.

The __drbd_chk_io_error() function uses begin_state_change_locked()
and end_state_change_locked(). That is only okay if the call site
holds a write-lock on state_rwlock.

Fixing that.