drbd: separate out receive function for P_OV_REPLY
This is preparation for synchronizing online verify with interval tree
and dagtags. We will keep objects to track online verify requests. Hence
P_OV_REPLY will refer to an existing object rather than initiating a new
request. This makes it sufficiently different that it no longer makes
sense to process it with receive_DataRequest.