ci: make release branch builds more generic
Branch master is now drbd-9.2, but we still want drbd-9.0 and drbd-9.1
builds. Use a more generic approach to avoid having to list all release
branches here.

Although this change is not really relevant on this branch, we apply the
changes here to keep the CI configuration as similar as possible between
branches.

Also combine the build:for-tests and build-latest jobs. They did almost
the same thing and this makes that clear. This has the effect that we no
longer re-build the ".$CI_COMMIT_SHA" versions on the release branch. We
have no need for that.

While we are at it, jump to the latest version of of build-helpers.