drbd: fix close-by-peer of control/meta socket
With a graceful disconnect, it might be that a node first notices
that the peer closed the control/meta socket before it processes the
P_TWOPC_COMMIT packet, that executes the disconnect locally.

In that specific case, do not go into the NETWORK_FAILURE state.

Note: In case the P_TWOPC_COMMIT packet is processed first, the
cstate is C_TEAR_DOWN. The state engine does not go from C_TEAR_DOWN
to C_NETWORK_FAILURE, it ignores that request for state change.