build,dkms: fix RDMA transport builds
in RC1 we saw that we can not assume that every kernel has
RDMA/INFINIBAND support (e.g, Ubuntu kvm kernels).

So make
- the build conditional
- make the dkms conditional as well

We have to test in the dkms as well, otherwise the build fails. Per se
module builds can not be made optional in dkms. The good news is that it
is just a shell script in the end.

Tested-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>