drbd: clear up resync requests from done_ee on disconnection
Suppose we are SyncTarget to a peer. We have sent a resync request,
received the reply and are writing the data. At this point we lose our
connection to the peer. Now drain_resync_activity waits for the resync
write to complete by waiting until sync_ee is empty. At this point, the
request is on done_ee.

This is normally harmless, because the request is cleared up when
drbd_finish_peer_reqs is called in conn_disconnect.

However, it causes problems when an application write from the peer
conflicted with the resync request.  This peer write is left on
active_ee and so conn_disconnect blocks waiting for active_ee to become
empty. This may result in warnings of the form:
INFO: task drbd_r_res:1234 blocked for more than 120 seconds.
And potentially similar for drbd_w_res as a knock-on effect.

Fix this with an additional call to drbd_finish_peer_reqs to clear up
any resync requests on done_ee.