containers: add flatcar-linux
This adds support for flatcar-linux. There is documentation on how to
build kernel modules for flatcar[1], but we don't need to do this dance
as all the necessary kernel headers/build bits are actually installed on
the host image anyways. So we can get away with a gentoo (stage3) image
which even already has all the build deps (make, gcc, curl).

For flatcar-linux we only support source builds, we don't provide
repos/binary packages.

While at it also clean up the Makefile:
Should be pretty obvious, the FORCE target looks weird, but that is how
to do it for pattern rules:
> Make does not consider implicit rules for PHONY targets.

It also removes the '-1' from the DRBD_VERSION (lbbuild does not use
that, so we never noticed).

[1]
https://www.flatcar-linux.org/docs/latest/reference/developer-guides/kernel-modules/