drbd: in case of discard_granularity do not send smaller resync requests
never send requests smaller than rs_discard_granularity. This makes
sure that we discard as much from the sync-target's backing device as
possible.

This change also introduces the idea that it can adjust the sleep time
slightly. The previous patch (round to power of 2) and this one
both allow that the actually requested number of sectors is different
from the number of sector that should be requested.
In order to resync with the requested rs_c_max_rate slightly adjust
the sleep time to correct for the rounding error of the actually
requested number of sectors.