drbd: avoid invoking make_resync_request() too early/often
Estimate the intervall in which we need to get the amount that
just came in to meet the c-max-rate.
We might call schedule make_resync_request() immediately,
schdule it earlier as planned, or let the timer expire, but
never stretch the cycle longer than RS_MAKE_REQS_INTV.

Please note the change to the controller function as well.
Assume that the next cycle will be RS_MAKE_REQS_INTV long.

TODO: This might become an issue for very high resync speeds