drbd: do not mark requests to diskless peers as RQ_NET_DONE
The event CONNECTION_LOST_WHILE_PENDING can occur for requests against
diskless peers. This caused RQ_NET_DONE to be set, which is confusing.
Mostly it went unnoticed. However, it prevented this fix from working in
some circumstances:
abef05ce0395 drbd: do not attempt to "resend" requests to diskless peers

Specifically this can cause problems when the connections from node A to
both peers B and C are lost. B with disk, and C diskless.
* Connection A->C lost. CONNECTION_LOST_WHILE_PENDING triggered.
  Requests marked with RQ_NET_DONE to diskless peer C.
* Connection A->B lost before these requests can complete.
* Connection A->C restored. RESEND triggered. Due to the RQ_NET_DONE
  bit, DRBD attempts to "resend" the requests to C.
* These requests get stuck because C never responds to them.