drbd: do not decrease the request size in an out-of-sync area
It is quite likely that that controller requests some BM_BLOCKS
in a cycle that is not a clean devisor of the current request size.
Make sure that within a cycle the request size is only allowed to
go up or stay the same. With that, we avoid that it appends a small
request at the end of a cycle. Then in the next cycle, it would need
to works its way up towards bigger sizes again.