debian: add drbd-ueficert package
Picked from 746952b5e9d9. This commit was applied to the drbd-9.1 branch
because drbd-9.0 was announced to be end-of-life after drbd-9.0.32.
However, several commits have been made to drbd-9.0 since, raising the
possibility of another release. Pick this build change to avoid
confusion should that happen.

Original commit message:

So far we shipped the secure boot key as part of the binary module. As a
file can only be provided by one package this can result in problems at
install time if there is already a module package installed for another
kernel.

Resolve this by shipping the key in an extra package and add a
dependency for it.

On key rollover we have to add a versioned dependency.