debian: add drbd-ueficert package
So far we shipped the secure boot key as part of the binary module. As a
file can only be provided by one package this can result in problems at
install time if there is already a module package installed for another
kernel.

Resolve this by shipping the key in an extra package and add a
dependency for it.

On key rollover we have to add a versioned dependency.