drbd: remove outdated comment from inc_ap_pending
The comment was outdated and misleading. For instance, the functions
tl_clear and tl_clear_barrier no longer exist. Since this count now
precisely corresponds to the number of requests with RQ_NET_PENDING,
just add that information directly to the documentation for
ap_pending_cnt and remove the otherwise unnecessary comment on
inc_ap_pending.