drbd: add RQ_UNMAP and RQ_ZEROES to RQ_STATE_0_MASK
These were already in local_rq_state. Add them to RQ_STATE_0_MASK so
that the comment there is correct. This has no functional effect.
RQ_STATE_0_MASK is only used in mod_rq_state, and that is never called
with RQ_UNMAP or RQ_ZEROES amongst the bits in the arguments.