drbd: Never scale down the sect_in
In the rare case that a cycle of generating resync requests took longer
than expected, do not scale down sect_in. They came in.

We scale it up in case the cycle is shortened, so that we can "ramp up"
resync speed quicker if we have a system of network link and data-providing
peer that have react quick to our requests.