drbd: kickstart the controller
The resync controller needs to be conservative with the first request.
It might be that the system of network and remote IO subsystem has
a very high latency. It would disturb DRBD's operation considerably
the resync process creates high congestion on the data socket immediately
at the start of a resync.

In case the first requests are completed within the first cycle we
want to kickstart the controller, as the latency of the to-be-controlled-system
("Regestrecke" in German language) is even lower than the resolution of
the digital control loop.