drbd: Improve drbd_get_listener()
The old implementation had the problem that multiple receivers thread
noticed that no listener exists concurrently. Then multiple receiver
threads try to create it concurrently. One succeeds and registers on
the list.  Under very special timings one receiver thread might return
EADDRINUSE. To mitigate that there was a retry loop around it. Good
enough that it never failed in our testing environment, others managed
to see the EADDRINUSE.

Improve that in a way that only one receiver thread creates the
listener and the other threads wait until that first thread finished
creating it.