drbd: Make demotion a two-phase commit; write bitmap during demote
One might expect that after a demote, the meta-data is stable. For
instance, the user may take a snapshot right after writing to the
resource on some other node. So far DRBD did not fulfill this
expectation since a demote was simply done locally and the peers got
informed about it.  They would then write the bitmap out after the state
change.

With this change, a demotion terminates after all peers wrote the bitmap.

With this patch, DRBD uses a two-phase commit to spread the news about the
demotion instead of a P_STATE packet. This does not require a new protocol
version, as no new packet gets introduced. Older versions process the
state change in the right way, the only difference being that they do
not write out the bitmap within the state changing two-phase commit.