drbd: skip done requests in tl_walk for connection
Since we perform tl_walk from finish_state_change(), we want to avoid
extremely long transfer log walks.

This situation could occur when we also have a lagging protocol A peer.

This is compensation for the fact that we no longer use
TL_NEXT_REQUEST_RESEND to avoid long transfer log walks under a lock.

In order to implement this, we reinterpret req_not_net_done slightly. It
can now point to requests which have not yet been sent.