drbd: cleaning up side-effects of supporting multiple diskless primaries
Allowing multiple diskless primaries creates some headaches. DRBD propagates
a new current-UUID in an interesting way. A receiving node needs to filter
out when should accept a new current-UUID and adopt it.
The most natural condition would be to role[NOW] == R_SECONDARY.
To handle multiply diskless primaries, that are not directly connected,
only indirectly via diskfull nodes it was necessary to weaken that
condition to also accept it on a primary already has the NEW_CUR_UUID
flag bit set. (Commit f550237a9c
'drbd: Fixup to the conditions when a diskless primary accepts a current UUID')

Only allow it if multiple primaries are allowed by the config.