drbd: Do not clear NEW_CUR_UUID when quorum comes back
This reverts a part of c4977df576
'drbd: consider that joining nodes might restore quorum after IO was suspended'

The idea of clearing NEW_CUR_UUID is wrong. It might be set due to events
before quorum was lost. Unfortunately, the commit message of that commit
is vague on the reasons, if there is a case where we want to avoid the resync
we need to avoid it by not setting the flag. Clearing it is wrong.