Revert "drbd: introduce caching pointer req_not_complete"
This reverts commit 606be5a4b8a62b736b15a7b01d6b28c232278b13.

This commit was added in order to avoid introducing a possible
performance regression. That related to the time spent iterating through
the transfer log while holding a spinlock.

Unfortunately, it caused a potential crash. Concurrent completion of
requests could have led to req_not_complete being set to a request which
is already complete. This request might then be destroyed, leaving
req_not_complete pointing to freed memory.

The race would have occurred as follows with tasks A and B:
* req_not_complete is set to request X
* A enters drbd_req_complete(X)
* A calls advance_req_not_complete(), finds request Y that is not yet
  complete
* B enters drbd_req_complete(Y)
* B calls advance_req_not_complete(), which has no effect because
  req_not_complete is still set to request X
* A sets req_not_complete to request Y

Since the benefit of this commit was a just a theoretical performance
improvement, revert the commit to avoid the problematic race.