drbd: change the return value of *_handshake functions
By returning the strategy and doing the strategy_to_repl_state() at the
call-site we get rid of the ugly overloading of the return type with
the values -1 (for reconnect) and -2 (for disconnect).

This is a preparation step for the next commit. We need to know the
calculated repl_state from the handshake, although the handshake
wants to abort an attempt to establish a connection.