drbd: unfreeze IO with an error if data gets lost while it lacks a quorum
Consider a configuration of
  on-no-quorum suspend-io;
  on-no-data-accessible io-error;
  rr-conflict retry;

When a primary loses quorum, IO freezes there. Now, in the other partition
a node gets promoted to a new primary, and the other partition creates a
new data generation.

When the IO-frozen primary tries to reconnect to the larger partition, the
attempt fails because two partitions with a primary can not join. But with
that, it learns that it no longer has the latest up-to-date data. Therefore it
completes the frozen IO requests with errors.
Ideally, that causes the application on top to terminate and when user-space
is cooperative it will even demote to secondary. (Containerized application
terminates, causing the "persistent volume" to be unmounted -> drbd demotes to
secondary)
then at the next connect attempt it can connect to the larger partition,
becomes resync-target. Everyone is happy about a complete automatic recovery
after a primary got isolated.