drbd: fix deadlock on demote while suspend-io is active
To avoid a sequence of
  open and auto-promote
  write to buffer cache
  close and demote
  system write-out of dirty pages to already demoted DRBD, changes are lost

we called fsync_bdev() on ourselves during demotion.
Which will seriously deadlock if we are currently in some "suspend-io" state.

Move the fsync_bdev() into drbd_release() instead.
The last one to close DRBD is responsible for the write-out of all dirty pages.

If that blocks, it blocks.
But we now can still reconfigure for io-error to unblock, if that is desired.