drbd: Clear NEW_CURRENT_UUID bit upon demote
The NEW_CURRENT_UUID flag bit gets set when the device loses contact to one of
the replicas (and the node is primary).

Its purpose is that just before the next change gets applied to the data-set
a new current-UUID gets generated. Usually, the next change is a write operation.
Also the decision to unfreeze IO and hence potentially unlocking a data-set
changing operation is one another possible reason.

When a node gives up its primary role, when it demotes to secondary, this
bit should be cleared, since the next IO unfreeze should not create a
new current-UUID. With the next promote to primary everything is re-evaluated
and the bit gets set again, it appropriate.