Revert "drbd: only create new current-uuid upon unfreezeing if we keep the requests"
This reverts commit 8bc0a9da10f1bfe4c13c66bf17cdb48a69063401.

It was just like an optimization to avoid the resync in that specific
scenario. It turns out that this was an overzealous optimization that
requires another exception where the new UUID is necessary, adding more code
and making everything more complex.

DRBD already contains code to automatically resolve this kind of split-brain
situation.  As we need that anyways, rely on that more often.
See commit d987c27a74a44d
'drbd: Automatically resolve split-brain in a specific case with quorum'

Co-developed-by: Joel Colledge <joel.colledge@linbit.com>