drbd: check presented UUID during connecting on a diskless primary node
This is another example of where the setting "on-no-data-accessible suspend-io"
was incomplete. (Well also applies if the diskless primaries IO was frozen
due to lack of quorum.)

When such a node connects to a node with a disk, it can only proceed with
accepting the connection when the diskful peer has the same current-uuid
as the expose_uuid of the diskless primary node. Otherwise, it would
"switch" the data generation(version) beneath the application.