drbd: introduce fail_io state member to abort all pending and future IO
Set it at the moment a diskless primary node that has IO frozen due to
lack of quorum, and it learns that the peers moved on to a newer data
generation.

Clear it when all openers closed the device

While the bit is set, no new openers are allowed.

Under the assumption that DRBD was auto-promoted, there is hope that with
ending all IO with IO-error, the opening applications exit or at least
close the DRBD device.

That allows a diskless node to adopt the current UUID of the peers that
moved on to a newer data generation.

Note, this fail_io member also enables a future implementation of a --force
flag for demoting a drbd device.