drbd: also generate bitmap_uuid for intentional diskless nodes
just throw them away when we reconnect.

This is reverts commit 0af5803b13fe31 from October 2016.

What was not considered back then is, that a diskless primary node might simply
freeze IO while it is disconnected from the nodes with storage. When in the
meantime, one of the nodes with storage gets promoted, and a new
data-generation gets generated in that process, DRBD will create a new
current-uuid.
This is necessary to detect if the nodes with disk moved on with the data
(created a new data generation) and if a reconnect is possible for a
IO suspended diskless primary or not.

Between October 2016 and now, it did not.

This changes the behavior regarding absent diskless nodes. Of course
we never want to have a UUID in the bitmap slot for connected diskless nodes.