drbd: ensure the exposed_data_uuid is set at the moment it might get relevance
We have the concept of delaying a new-current-uuid just before the first
write happens after a promote event. We apply the same principle to
the exposed_data_uuid (which is the volatile equivalent of the current-uuid,
on a diskless node).

As it is volatile by nature, it is often not on a previous value, but simply
all zero. Deriving the right value, just before we become a primary diskless
node from, either the detaching disk or the peers providing data.