drbd: update effective_size in time
With the next commit, a node might consider itself D_UP_TO_DATE earlier,
as the next commit ensures that the exposed current UUID always has a
value when it should have a value.

That would uncover a bug with the size determination code. Fix it here
before the next commit.