Revert "drbd: Only send early barriers when epoch is definitely finished"
This reverts commit 7f599c2606017ed6a5e4b9949f76b7cdb5290046.

This fix caused problems when a peer disk fails while application writes
are constantly pending. In this situation, the barrier for the final
writes to the peer with the failed disk was never sent, because
ap_write_cnt_total() was positive on every iteration.

Without this final barrier, the number of requests that did not yet
reach drbd_req_destroy() increases until something blocks. For instance,
if the Primary node is diskless, the activity log on another peer can
fill up until it cannot accept any more writes. Then this other
connection times out, as shown in this log:

[   35.353535] drbd res/0 drbd1 d-150.test: pdsk( UpToDate -> Failed )
[   35.356392] drbd res/0 drbd1: sending new current UUID: C7F9314BBEFB12F5
[   35.413470] drbd res/0 drbd1 d-150.test: pdsk( Failed -> Diskless )
[   37.344225] drbd res/0 drbd1 d-150.test: We did not send a P_BARRIER for 2013ms > ko-count (1) * timeout (20 * 0.1s); drbd kernel thread blocked?
[   37.406018] drbd res/0 drbd1 d-150.test: We did not send a P_BARRIER for 2075ms > ko-count (1) * timeout (20 * 0.1s); drbd kernel thread blocked?
[   37.412212] drbd res/0 drbd1 d-151.test: Remote failed to finish a request within 2001ms > ko-count (1) * timeout (20 * 0.1s)
[   37.417155] drbd res/0 drbd1 d-151.test: Request at 1012360s+4096 timed out
[   37.420745] drbd res/0 drbd1: Would lose quorum, but using tiebreaker logic to keep
[   37.424607] drbd res d-151.test: conn( Connected -> Timeout ) peer( Secondary -> Unknown )