drbd: send early barriers only when sender has seen all requests
This fixes the same problem as
7f599c260601 drbd: Only send early barriers when epoch is definitely finished
but without the problem that necessitated the revert in the previous
commit.

To quote the original fix:
This fixes a bug where 2 barriers were sent for the same epoch. A
request was in the epoch but not yet queued for sending. Hence the
sender thought that it had nothing to do and sent a barrier. Then the
request was processed and also caused a barrier to be sent.