drbd: Do not try reconciliation resync for diskless nodes
It is pointless to send the P_PEER_DAGTAG packet when one side is
diskless. Upon a primary node leaves a cluster unexpectedly they remaining
nodes use that packet to trigger a reconciliation resync.

It turned out that sending the current UUID from a diskless node that
promotes to primary can lead to wrong resync decissions later on in rare
cases.