compat: build compat tests with -Wno-error
Since Linux 5.15, "-Werror" is enabled on builds by default. Since we
use the Kbuild infrastructure to test against kernel features for compat,
that also applies to those builds.

The problem is that all compat tests produce warnings, so with >5.15
they all give false negative results.

So, (for now,) negate "-Werror" for the compat tests so that warnings
are no longer treated as errors.