compat: reverse genl_policy_in_ops patch
Compat patches should always be applied for the older kernels -- so
they should add back features that were removed in newer versions. This
one was the other way around: it applied to newer kernels, patching
"forward".

Reverse this so that (ideally) no compat patches are applied for the
latest kernel version.