drbd: prevent the resync requests being too large due to overflow
At multiples of 8TiB, a resync request could be sent that was larger
than DRBD's max bio size. This caused errors on the peer of the form:
drbd res/0 drbd1 d-151.test: /drbd/drbd/drbd_receiver.c:3216: sector: 17179869184s, size: 134217728

The issue originated in optimal_bits_for_alignment(). When the lower 32
bits of "bit" are 0x80000000, the expression "1 << __ffs(bit)" is
interpreted as the negative int 0x80000000, causing the function to
return this negative int. As a result "optimal_bits" had value
0x7fffffff and so a large request was sent.

Fix this by only considering the relevant bits of "bit".