drbd: Send ping with two-phase-commit prepare
There are many good reasons to finish two-phase-commit state changes as
quickly as possible.

Before this change, P_TWOPC_PREPARE (and P_TWOPC_PREP_RSZ) gets just sent via
the data socket. In case the connection to the peer gets interrupted it might
take up to a whole "ping-int" until we send a ping on that connection and
realize that the peer is no longer responsible.

Improve that by sending a ping on the control socket when we prepare
a two-phase-commit. With that we can abort the two-phase-commit promptly,
should one of the peers no longer respond.