Revert "drbd: Drop wire compatibility with versions prior to DRBD 9"
This reverts commit e00c2cc4571b7e22da0363b17e60bc198866996c, restoring
DRBD protocol compatibility back to version 86. That is, compatibility
with DRBD 8.0 to 8.4 is restored.

The motivation for this is to allow an upgrade path from 8.4 to 9.1
without requiring 9.0. This includes upgrades from Linux in-tree DRBD
8.4, as well as from the out-of-tree module. With this change,
maintenance of DRBD 9.0 is no longer necessary.

Since the planned changes which motivated e00c2cc4571b have not yet been
implemented, this revert does not cause any fundamental problems.