drbd: do not disconnect when no strategy determined from state
DRBD 8.4 peers may send P_STATE packets which we can actually ignore.
That is, we do not "consider_resync" and the peer is not negotiating an
attach. The commit
9ccb315bf916 drbd: change the return value of *_handshake functions
caused an undetermined strategy to be treated the same as a disconnect
strategy. As a result, this breaks compatibility with DRBD 8.4 peers.

Revert to treating an undetermined strategy as a no-op, as it was prior
to the above mentioned commit.