drbd: do not call sleeping send_buffer_half_full while holding spinlock
The function send_buffer_half_full() calls mutex_lock(). Hence we must
not call it while holding a spinlock. Avoid this by calling
send_buffer_half_full() before taking the spinlock.

This means that we no longer check the send buffer on every iteration.
However, this is not a fundamental change, because it could already be
filled up by the sender concurrently between checking it and making the
resync request. In other words, this check is just an optimization
heuristic.

This situation arose through some unfortunate merging side-effects.

Fixes: 20053a57999d ("Merge drbd-9.1/228f03577 drbd: refactoring make_resync_request() to improve readability")