drbd: do not drain resync writes with 8.4 peers when receiving write
This compatibility code is not necessary. Remove it.

This code has a convoluted history. It was introduced for the general
case, not just compatibility, in:
9d73d8559191 drbd: Fix a potential write ordering issue on SyncTarget nodes

Then DRBD started always using the AL on secondary nodes:
4a58f869e862 drbd: Always use the activity log when writing on a secondary node

As a result, this code was removed in:
47ec488ecb2d drbd: Remove redundant code

But then the logic was changed so that the AL was not exclusive with the
resync extents in:
4dc38cdeceea drbd: Break resync deadlock

And so the code was added back in:
524c6b7a7464 Revert "drbd: Remove redundant code"

Since then the AL has again been made exclusive with the resync extents:
ddc742b4beb7 drbd: drop special-casing peer_device in al_begin_io_for_peer

And this code was made conditional only for compatibility:
0b220db6109e drbd: move {active,read,sync,done}_ee lists from device to connection

But it is actually no longer necessary even for compatibility, for the
same reason described when it was first removed in 47ec488ecb2d. "In
DRBD9 every write on a secondary gets an entry in the activity log.
Therefore no explict check for a overlapping resync write is necessary
since, the call to drbd_al_begin_io() will block anyways until all
resync IO is done in that area."