compat: fix matching of wire_flags_to_bio_op
This fixes compat for 4.9.212-36.el7 and similar kernels. Without this
change, the error was output:
ERROR: A rule making an essential change was not executed!
ERROR: This would not show up as a compiler error, but would still break DRBD.
ERROR: As a precaution, the build will be aborted here.

This has been broken since
f60da60cc3d4 Revert "drbd: Drop wire compatibility with versions prior to DRBD 9"

This commit changed the relevant statement from a return to a variable
declaration and assignment. As a result, the compat patch did not match.
Fix this by matching the expression rather than the whole statement.

Also print the error to stderr instead of stdout.