drbd: respect peer writes held due to conflicts when barrier received
Prior to this commit, we could ack a barrier even though preceeding
writes were still held due to conflicts. In particular, the writes had
not completed locally. This can lead to data corruption when subsequent
writes arrive and are submitted in the wrong order.

Fix this by incrementing active_ee_cnt earlier. This ensures that
barriers also wait for requests which are held due to conflicts.

It is necessary to wait for free AL extents before incrementing
active_ee_cnt.

As part of the refactoring, remove the special case for compatibility
with 8.4 peers. This is not relevant because we no longer have resync
extents that are exclusive with activity log extents. It is fine to
activate AL extents on a Secondary even with a DRBD 8.4 peer.